requirements.txt
tkinterdnd2==0.4.3
python-docx==1.2.0
tqdm==4.67.1
openpyxl==3.1.5
